+++ src/css/ui/index.css
@@ -1,3 +1,3 @@
-@import url(/fe-kakaopage/src/css/ui/search-form.css);
-@import url(/fe-kakaopage/src/css/ui/btn.css);
-@import url(/fe-kakaopage/src/css/ui/lnb.css);
+@import url(/fe-kakaopage/src/css/ui/modules/search-form.css);
+@import url(/fe-kakaopage/src/css/ui/modules/btn.css);
+@import url(/fe-kakaopage/src/css/ui/modules/lnb.css);
