+++ src/js/sections/index.js
@@ -1,0 +1,3 @@
+export { initDailySection } from './modules/daily.js';
+export { initFooter } from './modules/footer.js';
+export { initGenreSections } from './modules/genre.js';
