+++ src/js/template/index.js
@@ -1,0 +1,1 @@
+export { gridItem } from './modules/gridItem.js';
